third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/border-color-expected.txt

This is a testharness.js-based test.
Found 4 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Can set 'border-top-color' to the 'currentcolor' keyword: currentcolor
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-left-color' to the 'currentcolor' keyword: currentcolor
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-right-color' to the 'currentcolor' keyword: currentcolor
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-bottom-color' to the 'currentcolor' keyword: currentcolor
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
Harness: the test ran to completion.
